feat: handle button events for calculation
buttonClickHandler(event) - determines which button
was pressed and takes an appropriate action.
clear() - Helper function invoked on 'AC' (clear)
action on calculator.
formatOutput (numStr) - used to format the number
output before printing to the screen to ensure
it's represented within the 11 character limit.
removeTrailingZeros(numStr) - a helper function
to deal with trailing zeros.